Implemented oid setter step for behave
