--- enginecore/features/steps/test_helpers.py
@@ -1,37 +1,0 @@
-"""A collection of shared utils for BDD tests"""
-import json
-
-
-from pysnmp import hlapi
-import websocket
-from circuits import Component
-
-
-def query_snmp_interface(oid, host="localhost", port=1024):
-    """Helper function to query snmp interface of a device"""
-    error_indicator, error_status, error_idx, var_binds = next(
-        hlapi.getCmd(
-            hlapi.SnmpEngine(),
-            hlapi.CommunityData("private", mpModel=0),
-            hlapi.UdpTransportTarget((host, port)),
-            hlapi.ContextData(),
-            hlapi.ObjectType(hlapi.ObjectIdentity(oid)),
-            lookupMib=False,
-        )
-    )
-
-    if error_indicator:
-        print(error_indicator)
-    elif error_status:
-        print(
-            "%s at %s"
-            % (
-                error_status.prettyPrint(),
-                error_idx and var_binds[int(error_idx) - 1][0] or "?",
-            )
-        )
-    else:
-        v_bind = var_binds[0]
-        return v_bind[1]
-
-    return None
